+++ react/OrderForm.tsx
@@ -6,24 +6,29 @@ import React, {
   useEffect,
   FC,
 } from 'react'
-import { useQuery } from 'react-apollo'
+import { ApolloError, useQuery } from 'react-apollo'
 
 import { orderForm as OrderFormQuery } from 'vtex.checkout-resources/Queries'
 
-import { dummyOrderForm } from './utils/dummyOrderForm'
+import { dummyOrderForm, emptyOrderForm } from './utils/dummyOrderForm'
+import { logSplunk } from './utils/logger'
 
 interface Context {
   loading: boolean
   orderForm: OrderForm | undefined
   setOrderForm: (orderForm: Partial<OrderForm>) => void
+  error: ApolloError | undefined
 }
 
 const OrderFormContext = createContext<Context | undefined>(undefined)
 
 export const OrderFormProvider: FC = ({ children }) => {
-  const { loading, data } = useQuery<{ orderForm: OrderForm }>(OrderFormQuery, {
-    ssr: false,
-  })
+  const { loading, data, error } = useQuery<{ orderForm: OrderForm }>(
+    OrderFormQuery,
+    {
+      ssr: false,
+    }
+  )
 
   const [orderForm, setOrderForm] = useReducer(
     (orderForm: OrderForm, newOrderForm: Partial<OrderForm>) => ({
@@ -34,21 +39,35 @@ export const OrderFormProvider: FC = ({ children }) => {
   )
 
   useEffect(() => {
+    if (error) {
+      logSplunk({
+        level: 'Important',
+        type: 'Error',
+        workflowType: 'OrderManager',
+        workflowInstance: 'orderform-query',
+        event: {
+          message: error.message,
+        },
+      })
+      console.error(error.message)
+    }
+
     if (loading) {
       return
     }
+
     data && setOrderForm(data.orderForm)
-  }, [data, loading])
+  }, [data, error, loading])
 
   const value = useMemo(
     () => ({
+      error,
       loading,
-      orderForm,
+      orderForm: error ? emptyOrderForm : orderForm,
       setOrderForm,
     }),
-    [loading, orderForm]
+    [error, loading, orderForm]
   )
-
   return (
     <OrderFormContext.Provider value={value}>
       {children}